refactor(routes): fold /drones templates+profiles+artifacts into drones.js (complete the module)
Folds the remaining /drones routes (job templates, compatibility check, profiles, artifacts —
including POST /drones/artifacts's requireAuth raw handler) into drones.js, so the module owns
ALL 32 /drones routes. mycelium.js now 5930 (−605 from drones total). Route contract unchanged
(284 routes), 0 /drones left inline, vitest green. Lucy (GLM-5.2), byte-reviewed by m5Max
(content-based signature match + gate).
