+++ server/routes/mycelium.js
@@ -4740,288 +4740,9 @@ registerDroneRoutes(router, {
   asyncHandler, checkAgentOrAdmin, checkAgent, checkAdmin, checkGuardrails,
   agentWriteLimiter, escapeHtml, parseLimit, parseIntParam, validateEnum,
   apiError, emitEvent, getAdminDisplayName, isAdminKey, getStudioUser,
+  requireAuth, artifactUpload, ARTIFACTS_DIR,
   DRONE_JOB_STATUSES,
 });
-
-// ======== JOB TEMPLATES (reusable job type definitions for smart routing) ========
-// Must be before /drones/:id to prevent :id from catching "templates"
-
-router.get('/drones/templates', asyncHandler(function (req, res) {
-  var who = checkAgentOrAdmin(req, res);
-  if (!who) return;
-  res.json(listJobTemplates());
-}));
-
-router.get('/drones/templates/:id', asyncHandler(function (req, res) {
-  var who = checkAgentOrAdmin(req, res);
-  if (!who) return;
-  var template = getJobTemplate(req.params.id);
-  if (!template) return res.status(404).json({ error: 'Template not found' });
-  res.json(template);
-}));
-
-router.post('/drones/templates', asyncHandler(function (req, res) {
-  if (!checkAdmin(req, res)) return;
-  var id = req.body.id;
-  if (!id) return res.status(400).json({ error: 'id is required' });
-  if (getJobTemplate(id)) return res.status(409).json({ error: 'Template already exists: ' + id });
-  var template = createJobTemplate(id, req.body);
-  emitEvent('job_template_created', getAdminDisplayName(req), 'drone', 'Created job template: ' + id);
-  res.json(template);
-}));
-
-router.put('/drones/templates/:id', asyncHandler(function (req, res) {
-  if (!checkAdmin(req, res)) return;
-  var template = getJobTemplate(req.params.id);
-  if (!template) return res.status(404).json({ error: 'Template not found' });
-  var updated = updateJobTemplate(req.params.id, req.body);
-  emitEvent('job_template_updated', getAdminDisplayName(req), 'drone', 'Updated job template: ' + req.params.id);
-  res.json(updated);
-}));
-
-router.delete('/drones/templates/:id', asyncHandler(function (req, res) {
-  if (!checkAdmin(req, res)) return;
-  var template = getJobTemplate(req.params.id);
-  if (!template) return res.status(404).json({ error: 'Template not found' });
-  deleteJobTemplate(req.params.id);
-  emitEvent('job_template_deleted', getAdminDisplayName(req), 'drone', 'Deleted job template: ' + req.params.id);
-  res.json({ ok: true, deleted: req.params.id });
-}));
-
-// ======== DRONE COMPATIBILITY CHECK ========
-
-router.get('/drones/:id/compatibility', asyncHandler(function (req, res) {
-  var who = checkAgentOrAdmin(req, res);
-  if (!who) return;
-  var agent = getAgent(req.params.id);
-  if (!agent) return res.status(404).json({ error: 'Drone not found' });
-  var result = checkDroneCompatibility(req.params.id);
-  res.json(result);
-}));
-
-// ======== DRONE PROFILES (per-drone setup & dependency definitions) ========
-// Must be before /drones/:id to prevent :id from catching "profiles"
-
-// List all profiles
-router.get('/drones/profiles', asyncHandler(function (req, res) {
-  var who = checkAgentOrAdmin(req, res);
-  if (!who) return;
-  var profiles = listDroneProfiles();
-  profiles = profiles.map(function (p) {
-    p.requires = JSON.parse(p.requires || '{}');
-    p.artifacts = JSON.parse(p.artifacts || '[]');
-    p.env = JSON.parse(p.env || '{}');
-    p.drones = getDronesWithProfile(p.id);
-    return p;
-  });
-  res.json(profiles);
-}));
-
-// Create a profile (admin only)
-router.post('/drones/profiles', asyncHandler(function (req, res) {
-  if (!checkAdmin(req, res)) return;
-  var id = req.body.id;
-  if (!id) return res.status(400).json({ error: 'id is required (slug, e.g. "kc-art-gen")' });
-  if (!req.body.name) return res.status(400).json({ error: 'name is required' });
-  if (getDroneProfile(id)) return res.status(409).json({ error: 'Profile already exists: ' + id });
-  var profile = createDroneProfile(
-    id, req.body.name, req.body.description,
-    req.body.requires, req.body.artifacts,
-    req.body.setup_script, req.body.workspace, req.body.env
-  );
-  emitEvent('drone_profile_created', '__admin__', 'drone', 'Created drone profile: ' + id);
-  res.json(profile);
-}));
-
-// Get single profile
-router.get('/drones/profiles/:profileId', asyncHandler(function (req, res) {
-  var who = checkAgentOrAdmin(req, res);
-  if (!who) return;
-  var profile = getDroneProfile(req.params.profileId);
-  if (!profile) return res.status(404).json({ error: 'Profile not found' });
-  profile.requires = JSON.parse(profile.requires || '{}');
-  profile.artifacts = JSON.parse(profile.artifacts || '[]');
-  profile.env = JSON.parse(profile.env || '{}');
-  profile.drones = getDronesWithProfile(profile.id);
-  res.json(profile);
-}));
-
-// Update a profile (admin only) — invalidates setup_done for all assigned drones
-router.put('/drones/profiles/:profileId', asyncHandler(function (req, res) {
-  if (!checkAdmin(req, res)) return;
-  var profile = getDroneProfile(req.params.profileId);
-  if (!profile) return res.status(404).json({ error: 'Profile not found' });
-  var updated = updateDroneProfile(req.params.profileId, req.body);
-  updated.requires = JSON.parse(updated.requires || '{}');
-  updated.artifacts = JSON.parse(updated.artifacts || '[]');
-  updated.env = JSON.parse(updated.env || '{}');
-  emitEvent('drone_profile_updated', '__admin__', 'drone', 'Updated drone profile: ' + req.params.profileId + ' (all drone setups invalidated)');
-  res.json(updated);
-}));
-
-// Delete a profile (admin only)
-router.delete('/drones/profiles/:profileId', asyncHandler(function (req, res) {
-  if (!checkAdmin(req, res)) return;
-  var profile = getDroneProfile(req.params.profileId);
-  if (!profile) return res.status(404).json({ error: 'Profile not found' });
-  deleteDroneProfile(req.params.profileId);
-  emitEvent('drone_profile_deleted', '__admin__', 'drone', 'Deleted drone profile: ' + req.params.profileId);
-  res.json({ ok: true, deleted: req.params.profileId });
-}));
-
-// Assign profile to drone
-router.post('/drones/profiles/:profileId/assign', asyncHandler(function (req, res) {
-  if (!checkAdmin(req, res)) return;
-  var profileId = req.params.profileId;
-  var droneId = req.body.drone_id;
-  if (!droneId) return res.status(400).json({ error: 'drone_id is required' });
-  if (!getDroneProfile(profileId)) return res.status(404).json({ error: 'Profile not found: ' + profileId });
-  var agent = getAgent(droneId);
-  if (!agent) return res.status(404).json({ error: 'Drone not found: ' + droneId });
-  assignDroneProfile(droneId, profileId);
-  emitEvent('drone_profile_assigned', '__admin__', 'drone', 'Assigned profile ' + profileId + ' to drone ' + droneId);
-  res.json({ ok: true, drone_id: droneId, profile_id: profileId });
-}));
-
-// Unassign profile from drone
-router.delete('/drones/profiles/:profileId/assign/:droneId', asyncHandler(function (req, res) {
-  if (!checkAdmin(req, res)) return;
-  unassignDroneProfile(req.params.droneId, req.params.profileId);
-  res.json({ ok: true, unassigned: true });
-}));
-
-// Get profiles assigned to a drone (used by worker at boot)
-router.get('/drones/profiles/by-drone/:droneId', asyncHandler(function (req, res) {
-  var who = checkAgentOrAdmin(req, res);
-  if (!who) return;
-  var assignments = getDroneProfileAssignments(req.params.droneId);
-  assignments = assignments.map(function (a) {
-    a.requires = JSON.parse(a.requires || '{}');
-    a.artifacts = JSON.parse(a.artifacts || '[]');
-    a.env = JSON.parse(a.env || '{}');
-    return a;
-  });
-  res.json(assignments);
-}));
-
-// Drone reports profile setup complete
-router.post('/drones/profiles/:profileId/setup-complete', asyncHandler(function (req, res) {
-  var agentId = checkAgent(req, res);
-  if (!agentId) return;
-  var profileId = req.params.profileId;
-  if (!getDroneProfile(profileId)) return res.status(404).json({ error: 'Profile not found' });
-  var checksum = req.body.checksum || '';
-  markProfileSetupDone(agentId, profileId, checksum);
-  emitEvent('drone_profile_setup_done', agentId, 'drone', agentId + ' completed setup for profile ' + profileId);
-  res.json({ ok: true, profile_id: profileId, setup_done: true });
-}));
-
-// ======== DRONE ARTIFACTS (persistent files — models, LoRAs, etc.) ========
-// Must be before /drones/:id to prevent :id from catching "artifacts"
-
-// Upload a drone artifact (persistent, no TTL) — ADMIN ONLY.
-// Artifacts (e.g. generate_flux.py, model weights) are trusted files that drones
-// download and EXECUTE. The store overwrites by filename, so a non-admin upload of
-// name=generate_flux.py would poison the trusted script → mesh-wide RCE (C-3).
-// The admin gate MUST run before multer writes the file to disk, so it sits ahead
-// of artifactUpload in the chain rather than inside the handler.
-router.post('/drones/artifacts', requireAuth, function (req, res, next) {
-  var who = checkAgentOrAdmin(req, res);
-  if (!who) return; // 401 already sent
-  if (!req._authIsAdmin) return apiError(res, 403, 'Artifact upload is admin-only — artifacts are trusted code executed on drones.');
-  next();
-}, artifactUpload.single('file'), asyncHandler(function (req, res) {
-  var who = checkAgentOrAdmin(req, res);
-  if (!who) return;
-  if (!req.file) return res.status(400).json({ error: 'No file uploaded. Use multipart form with field name "file"' });
-  var protocol = req.headers['x-forwarded-proto'] || req.protocol;
-  var host = req.headers['x-forwarded-host'] || req.get('host');
-  var baseUrl = protocol + '://' + host;
-  var url = baseUrl + '/api/mycelium/drones/artifacts/' + req.file.filename;
-  var sizeStr = req.file.size > 1024 * 1024 ? Math.round(req.file.size / 1024 / 1024) + 'MB' : Math.round(req.file.size / 1024) + 'KB';
-  emitEvent('artifact_uploaded', who, 'drone', 'Uploaded drone artifact: ' + req.file.filename + ' (' + sizeStr + ')');
-  res.json({ ok: true, name: req.file.filename, url: url, size: req.file.size });
-}));
-
-// List drone artifacts
-router.get('/drones/artifacts', asyncHandler(function (req, res) {
-  var who = checkAgentOrAdmin(req, res);
-  if (!who) return;
-  var protocol = req.headers['x-forwarded-proto'] || req.protocol;
-  var host = req.headers['x-forwarded-host'] || req.get('host');
-  var baseUrl = protocol + '://' + host;
-  var artifacts = [];
-  try {
-    artifacts = fs.readdirSync(ARTIFACTS_DIR).map(function (f) {
-      var stat = fs.statSync(nodePath.join(ARTIFACTS_DIR, f));
-      return { name: f, size: stat.size, uploaded: stat.mtime.toISOString(), url: baseUrl + '/api/mycelium/drones/artifacts/' + f };
-    });
-  } catch (e) { /* empty */ }
-  res.json(artifacts);
-}));
-
-// Download a drone artifact (auth required)
-router.get('/drones/artifacts/:name', asyncHandler(function (req, res) {
-  if (!checkAgentOrAdmin(req, res)) return;
-  var name = req.params.name.replace(/[^a-zA-Z0-9_.\-]/g, '');
-  var filePath = nodePath.join(ARTIFACTS_DIR, name);
-  if (!fs.existsSync(filePath)) return res.status(404).json({ error: 'Artifact not found' });
-  res.download(filePath, name);
-}));
-
-// Delete a drone artifact (admin only)
-router.delete('/drones/artifacts/:name', asyncHandler(function (req, res) {
-  if (!checkAdmin(req, res)) return;
-  var name = req.params.name.replace(/[^a-zA-Z0-9_.\-]/g, '');
-  var filePath = nodePath.join(ARTIFACTS_DIR, name);
-  if (!fs.existsSync(filePath)) return res.status(404).json({ error: 'Artifact not found' });
-  fs.unlinkSync(filePath);
-  res.json({ ok: true, deleted: name });
-}));
-
-// Get single drone + recent jobs + diagnostics (must be after /drones/artifacts to prevent :id catching "artifacts")
-router.get('/drones/:id', asyncHandler(function (req, res) {
-  var who = checkAgentOrAdmin(req, res);
-  if (!who) return;
-  var agent = getAgent(req.params.id);
-  if (!agent || agent.project_id !== 'drone') return res.status(404).json({ error: 'Drone not found' });
-  var { api_key_hash, ...safe } = agent;
-  var recentJobs = listDroneJobs({ drone_id: req.params.id, limit: 20 });
-  safe.recent_jobs = recentJobs;
-  // Include system diagnostics from latest savepoint
-  var savepoint = getLatestSavepoint(req.params.id);
-  if (savepoint) {
-    try {
-      var snapshot = JSON.parse(savepoint.state_snapshot || '{}');
-      safe.system_info = snapshot.system_info || null;
-      safe.warnings = snapshot.warnings || [];
-      safe.worker_version = snapshot.worker_version || null;
-    } catch (e) { console.warn('[mycelium] JSON parse failed for state_snapshot (drone: ' + req.params.id + '):', e.message); }
-  }
-  // Error summary from recent failed jobs
-  var failedJobs = recentJobs.filter(function (j) { return j.status === 'failed'; });
-  if (failedJobs.length > 0) {
-    safe.error_summary = failedJobs.slice(0, 5).map(function (j) {
-      var resultData = {};
-      try { resultData = JSON.parse(j.result_data || '{}'); } catch (e) { console.warn('[mycelium] JSON parse failed for result_data (job: ' + j.id + '):', e.message); }
-      return {
-        job_id: j.id,
-        title: j.title,
-        error_type: resultData.error_type || 'unknown',
-        message: (resultData.message || j.error || '').substring(0, 200),
-        suggestion: (resultData.suggestion || '').substring(0, 200),
-        failed_at: j.completed_at,
-      };
-    });
-  }
-  // Include profile assignments
-  safe.profiles = getDroneProfileAssignments(req.params.id).map(function (a) {
-    a.requires = JSON.parse(a.requires || '{}');
-    a.artifacts = JSON.parse(a.artifacts || '[]');
-    return a;
-  });
-  res.json(safe);
-}));
 
 // =============== APPROVALS ===============
